@@ -1,6 +1,6 @@
 <?php
-
-use App\User;
+namespace Database\Seeders;
+use App\Models\User;
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Str;
 
@@ -21,7 +21,7 @@ class UsersTablesSeeder extends Seeder
             'password'          => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password
             'remember_token'    => Str::random(10),
         ])->assignRole('Admin');
-        
-        factory(User::class, 30)->create();
+
+        User::factory(100)->create();
     }
 }
