@@ -1,53 +1,55 @@
 <?php
 
-/** @var Factory $factory */
+namespace Database\Factories;
 
 use App\Models\BroadCategory;
-use App\Models\Complaint;
 use App\Models\District;
 use App\Models\Program;
 use App\Models\Project;
 use App\Models\Province;
 use App\Models\SpecificCategory;
-use App\User;
-use Faker\Generator as Faker;
-use Illuminate\Database\Eloquent\Factory;
+use App\Models\User;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| This directory should contain each of the model factory definitions for
-| your application. Factories provide a convenient way to generate new
-| model instances for testing / seeding your application's database.
-|
-*/
 
-$factory->define(Complaint::class, function (Faker $faker) {
-    $gender = ['Male', 'Female'];
-    $status = ['Registered', 'Under Investigation', 'Solved'];
-    $quarter = ['First Quarter', 'Second Quarter', 'Third Quarter','Fourth Quarter'];
+/**
+ * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\Complaint>
+ */
+class ComplaintFactory extends Factory
+{
 
-    return [
-        'caller_name'              => $faker->name,
-        'tel_no_received'          => $faker->phoneNumber,
-        'gender'                   => $gender[$faker->numberBetween(0, 1)],
-        'received_date'            => $faker->dateTimeBetween('-4 years', now()),
-        'status'                   => $status[$faker->numberBetween(0, 2)],
-        'Quarter'                  => $quarter[$faker->numberBetween(0, 2)],
-        'referred_to'              => $faker->name,
-        'beneficiary_file'         => null,
-        'broad_category_id'        => BroadCategory::all()->random()->id,
-        'specific_category_id'     => SpecificCategory::all()->random()->id,
-        'person_who_shared_action' => $faker->name,
-        'close_date'               => $faker->date('Y-m-d'),
-        'description'              => $faker->text(20),
-        'province_id'              => Province::all()->random()->id,
-        'district_id'              => District::all()->random()->id,
-        'village'                  => $faker->text(10),
-        'user_id'                  => User::all()->random()->id,
-        'project_id'               => Project::all()->random()->id,
-        'program_id'               => Program::all()->random()->id
-    ];
-});
+    /**
+     * Define the model's default state.
+     *
+     * @return array<string, mixed>
+     */
+    public function definition()
+    {
+        $gender = ['Male', 'Female'];
+        $status = ['Registered', 'Under Investigation', 'Solved'];
+        $quarter = ['First Quarter', 'Second Quarter', 'Third Quarter', 'Fourth Quarter'];
+
+        return [
+            'caller_name'              => fake()->name,
+            'tel_no_received'          => fake()->phoneNumber,
+            'gender'                   => $gender[fake()->numberBetween(0, 1)],
+            'received_date'            => fake()->dateTimeBetween('-4 years', now()),
+            'status'                   => $status[fake()->numberBetween(0, 2)],
+            'Quarter'                  => $quarter[fake()->numberBetween(0, 2)],
+            'referred_to'              => fake()->name,
+            'beneficiary_file'         => null,
+            'broad_category_id'        => BroadCategory::all()->random()->id,
+            'specific_category_id'     => SpecificCategory::all()->random()->id,
+            'person_who_shared_action' => fake()->name,
+            'close_date'               => fake()->date('Y-m-d'),
+            'description'              => fake()->text(20),
+            'province_id'              => Province::all()->random()->id,
+            'district_id'              => District::all()->random()->id,
+            'village'                  => fake()->text(10),
+            'user_id'                  => User::all()->random()->id,
+            'project_id'               => Project::all()->random()->id,
+            'program_id'               => Program::all()->random()->id
+        ];
+    }
+}
+
