@@ -15,16 +15,13 @@ class CreateActivityLogTable extends Migration
             $table->bigIncrements('id');
             $table->string('log_name')->nullable();
             $table->text('description');
-            $table->unsignedBigInteger('subject_id')->nullable();
-            $table->string('subject_type')->nullable();
-            $table->unsignedBigInteger('causer_id')->nullable();
-            $table->string('causer_type')->nullable();
+            $table->nullableMorphs('subject', 'subject');
+            $table->string('event')->nullable();
+            $table->uuid('batch_uuid')->nullable();
+            $table->nullableMorphs('causer', 'causer');
             $table->json('properties')->nullable();
             $table->timestamps();
-
             $table->index('log_name');
-            $table->index(['subject_id', 'subject_type'], 'subject');
-            $table->index(['causer_id', 'causer_type'], 'causer');
         });
     }
 
