@@ -1,6 +1,6 @@
 <?php
 
-
+namespace Database\Seeders;
 use App\Models\SpecificCategory;
 use Illuminate\Database\Seeder;
 
